implements a single benchmark, using criterion library. Required a pretty agressive reorganization of modules and src files
